requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
appnope==0.1.0
APScheduler==3.6.3
astroid==2.3.3
autopep8==1.5
backcall==0.1.0
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
decorator==4.4.2
distlib==0.3.0
filelock==3.0.12
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
forex-python==1.5
gunicorn==20.0.4
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
multitasking==0.0.9
numpy==1.18.4
pandas==1.0.4
pandas-datareader==0.8.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2020.5.28
prompt-toolkit==3.0.3
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.5.2
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
rope==0.16.0
schedule==0.6.0
simplejson==3.17.0
six==1.15.0
SQLAlchemy==1.3.17
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.9
virtualenv==20.0.21
virtualenv-clone==0.5.4
wcwidth==0.1.8
Werkzeug==1.0.1
wrapt==1.11.2
